move legacy start to point of use
