+++ pkg/authorization/apiserver/admission/restrictusers/restrictusers.go
@@ -20,7 +20,7 @@ import (
 	authorizationtypedclient "github.com/openshift/client-go/authorization/clientset/versioned/typed/authorization/v1"
 	userclient "github.com/openshift/client-go/user/clientset/versioned"
 	userinformer "github.com/openshift/client-go/user/informers/externalversions"
-	oadmission "github.com/openshift/origin/pkg/cmd/server/admission"
+	oadmission "github.com/openshift/origin/pkg/admission/admissioninitializers"
 	usercache "github.com/openshift/origin/pkg/user/cache"
 )
