--- pkg/cmd/server/election/doc.go
@@ -1,2 +1,0 @@
-// Package election provides objects for managing the list of active masters via leases.
-package election
